add_classes_to_represent_model (Generate Diabetes Assessment)
- added classes to hold Patient and PatientNote objects from database retrieval
- created PatientAssessment to hold a Patient, their notes, and their risk
- created Risk enum to hold risk assessment result

Used an enum to hold the assessment result to ensure the value always matched one of the set terms, while also allowing for additional granularity to be easily added in future.
In order to allow output to match requested format a custom method was added to convert this enum to a String (as the desired format includes spaces). Possible there may be a cleaner way to implement this later.

https://trello.com/c/cedtNn3j/7-generate-diabetes-assessment